// src/Types/Enums/enums.type.ts
export enum GenderEnum {
    MALE = 'male',
    FEMALE = 'female',
    NOT_SPECIFIED = 'not-specified'
}
export enum RolesEnum {
    USER = 'user',
    ADMIN = 'admin'
}
export enum ProviderEnum {
    SYSTEM = 'system',
    GOOGLE = 'google'
}
export enum OtpEnum {
    CONFIRM_EMAIL = "confirmEmail",
    FORGET_PASSWORD = "forgetPassword"
}
export enum ImageExtensions {
    JPG = 'jpg',
    JPEG = 'jpeg',
    PNG = 'png',
    GIF = 'gif',
    WEBP = 'webp',
    BMP = 'bmp',
    SVG = 'svg',
}

export enum LabelAddressEnum {
    HOME = 'home',
    WORK = 'work',
    OTHER = 'other'
}
export enum PaymentMethodsEnum {
    CASH = 'cash',
    CREDIT_CARD = 'credit_card'
}
export enum PaymentStatusEnum {
    PENDING = 'pending',
    PAID = 'paid',
    FAILED = 'failed'
}
export enum OrderStatusEnum {
    PENDING = 'pending',
    PLACED = 'placed',
    SHIPPED = 'shipped',
    DELIVERD = 'delivered',
    CANCELLD = 'cancelld',
    REFUNDET = 'refundet'
}

export enum CouponDurationEnum {
    ONCE = 'once',
    REPEATING = 'repeating',
    FOREVER = 'forever'

}